Move search text field out of its own package
